requirements.txt
Flask
python-dotenv
requests
psycopg2-binary
Flask-SQLAlchemy
flask-login
flask_wtf
postgresql
